+++ validate/fast-tests/100TestHEPEvt.sh
@@ -3,13 +3,31 @@
 # Write a ROOT file in the default geometry.  This output file should
 # be tested to make sure that all of the fields are readable.
 
-OUTPUT=100TestHEPEvt.root
+# Get the directory containing the script from the command line
+# parameters (avoids bash trickery).  Use the current directory as the
+# default.
+DIR=.
+if [ ${#1} -gt 0 ]; then
+    DIR=${1}
+fi
+
+# Make sure that edep-sim has been setup.
+if ! which edep-sim; then
+    echo FAIL: Executable not found for edep-sim
+    exit 1
+fi
+
+# Setup the base names 
+BASE=100TestHEPEvt
+
+OUTPUT=${BASE}.root
 
 if [ -f ${OUTPUT} ]; then
     rm ${OUTPUT}
 fi
 
-cat > 100TestHEPEvt.txt <<EOF
+INPUT=${BASE}.txt
+cat > ${INPUT} <<EOF
          5
    1        13    0    0 -0.21729988e+02 -0.16027278e+02  0.67855399e+02  0.10566000e+00
    1        22    0    0  0.25382032e-04  0.42110624e-03  0.14112842e-04  0.00000000e+00
@@ -24,7 +42,8 @@ cat > 100TestHEPEvt.txt <<EOF
    1        11    0    0  0.18675520e+01 -0.14315622e+02  0.24412754e+02  0.51000000e-03
 EOF
 
-cat > 100TestHEPEvt.mac <<EOF
+MACRO=${BASE}.mac
+cat > ${MACRO} <<EOF
 #######################################
 # Set the hit segment.
 #######################################
@@ -33,7 +52,7 @@ cat > 100TestHEPEvt.mac <<EOF
 /edep/hitLength drift 1.0 mm
 /edep/update
 
-/generator/kinematics/hepevt/input 100TestHEPEvt.txt
+/generator/kinematics/hepevt/input ${INPUT}
 /generator/kinematics/hepevt/verbose 2
 /generator/kinematics/set hepevt
 
@@ -42,4 +61,4 @@ cat > 100TestHEPEvt.mac <<EOF
 /generator/add
 EOF
 
-edep-sim -o ${OUTPUT} -C -e 3 100TestHEPEvt.mac || exit 1
+edep-sim -o ${OUTPUT} -C -e 3 ${MACRO} || exit 1
